requirements.txt
flask
pycryptodomex
pandas
requests
sklearn
joblib